pyproject.toml
[project]
name = "tradeops"
version = "0.1.0"
description = "CLI trade-ops copilot for Alpaca paper trading"
requires-python = ">=3.11"
dependencies = [
  "alpaca-py>=0.42.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "pyyaml>=6.0.1",
  "rich>=13.7.1",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
tradeops = "tradeops.app.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tradeops", "tradeops.app"]

[tool.pytest.ini_options]
testpaths = ["tests"]

